refact: 開くボタンをテストを別ファイルに切り出した
