+++ src/components/BookmarkForm.test.tsx
@@ -53,62 +53,6 @@ describe('BookmarkForm', () => {
     expect(urlElement).toHaveValue(targetBookmark.url)
   })
 
-  describe('開くボタンの動作', () => {
-    it('開くボタンをクリックしたときに、正しいURLと属性で window.open が呼び出されること', async () => {
-      const openSpy = vi.spyOn(window, 'open').mockImplementation(() => null)
-
-      const user = userEvent.setup()
-      const targetBookmark = TestBookmarks[0]
-
-      render(<BookmarkForm key={targetBookmark.id} bookmark={targetBookmark} />)
-
-      await clickButton(user, UI_LABELS.ACTIONS.OPEN)
-
-      expect(openSpy).toHaveBeenCalledTimes(1) // 1回だけ呼ばれたか
-      expect(openSpy).toHaveBeenCalledWith(
-        targetBookmark.url,
-        '_blank', // 第2引数：ターゲット
-        'noreferrer', // 第3引数：機能文字列
-      )
-
-      openSpy.mockRestore()
-    })
-
-    it('URLを変更した場合には、変更したURLでwindow.open が呼び出されること', async () => {
-      const openSpy = vi.spyOn(window, 'open').mockImplementation(() => null)
-
-      const user = userEvent.setup()
-      const targetBookmark = TestBookmarks[0]
-      const inputUrl = TestBookmarks[1].url
-
-      render(<BookmarkForm key={targetBookmark.id} bookmark={targetBookmark} />)
-
-      await inputText(user, UI_LABELS.FIELDS.URL, inputUrl)
-      await clickButton(user, UI_LABELS.ACTIONS.OPEN)
-
-      expect(openSpy).toHaveBeenCalledTimes(1) // 1回だけ呼ばれたか
-      expect(openSpy).toHaveBeenCalledWith(
-        inputUrl,
-        '_blank', // 第2引数：ターゲット
-        'noreferrer', // 第3引数：機能文字列
-      )
-
-      openSpy.mockRestore()
-    })
-
-    it('URLのテキストボックスに不正な文字列が入力されている場合には開くボタンは無効となること', async () => {
-      const user = userEvent.setup()
-      const targetBookmark = TestBookmarks[0]
-
-      render(<BookmarkForm key={targetBookmark.id} bookmark={targetBookmark} />)
-
-      await inputText(user, UI_LABELS.FIELDS.URL, INVALID_STRING.URL)
-      const openButton = await clickButton(user, UI_LABELS.ACTIONS.OPEN)
-
-      expect(openButton).toBeDisabled()
-    })
-  })
-
   describe('戻るボタンの動作', () => {
     beforeEach(() => {
       vi.resetAllMocks()
